Move wasp-cli into own go submodule
