+++ tools/wasp-cli/authentication/cmd.go
@@ -3,7 +3,7 @@ package authentication
 import (
 	"github.com/spf13/cobra"
 
-	"github.com/iotaledger/wasp/tools/wasp-cli/log"
+	"github.com/iotaledger/wasp/packages/log"
 )
 
 var authCmd = &cobra.Command{
