+++ tools/wasp-cli/chain/activate.go
@@ -5,8 +5,8 @@ import (
 
 	"github.com/spf13/cobra"
 
+	"github.com/iotaledger/wasp/packages/log"
 	"github.com/iotaledger/wasp/packages/registry"
-	"github.com/iotaledger/wasp/tools/wasp-cli/log"
 )
 
 var HTTP404ErrRegexp = regexp.MustCompile(`"Code":404`)
