+++ tools/wasp-cli/chain/evm.go
@@ -10,9 +10,9 @@ import (
 	"github.com/spf13/cobra"
 
 	"github.com/iotaledger/wasp/packages/evm/evmtypes"
+	"github.com/iotaledger/wasp/packages/log"
 	"github.com/iotaledger/wasp/packages/util"
 	"github.com/iotaledger/wasp/packages/vm/core/evm"
-	"github.com/iotaledger/wasp/tools/wasp-cli/log"
 )
 
 type evmDeployParams struct {
