+++ tools/wasp-cli/chain/governance.go
@@ -8,8 +8,8 @@ import (
 
 	"github.com/iotaledger/wasp/client/chainclient"
 	"github.com/iotaledger/wasp/packages/cryptolib"
+	"github.com/iotaledger/wasp/packages/log"
 	"github.com/iotaledger/wasp/packages/vm/core/governance"
-	"github.com/iotaledger/wasp/tools/wasp-cli/log"
 )
 
 func changeAccessNodesCmd() *cobra.Command {
