+++ tools/wasp-cli/config/config.go
@@ -9,9 +9,9 @@ import (
 
 	"github.com/iotaledger/wasp/client"
 	"github.com/iotaledger/wasp/packages/l1connection"
+	"github.com/iotaledger/wasp/packages/log"
 	"github.com/iotaledger/wasp/packages/parameters"
 	"github.com/iotaledger/wasp/packages/testutil/privtangle/privtangledefaults"
-	"github.com/iotaledger/wasp/tools/wasp-cli/log"
 )
 
 var (
