+++ tools/wasp-cli/metrics/consensus.go
@@ -8,8 +8,8 @@ import (
 
 	iotago "github.com/iotaledger/iota.go/v3"
 	"github.com/iotaledger/wasp/packages/isc"
+	"github.com/iotaledger/wasp/packages/log"
 	"github.com/iotaledger/wasp/tools/wasp-cli/config"
-	"github.com/iotaledger/wasp/tools/wasp-cli/log"
 )
 
 var timestampNeverConst = time.Time{}
