+++ tools/wasp-cli/metrics/metrics.go
@@ -3,7 +3,7 @@ package metrics
 import (
 	"github.com/spf13/cobra"
 
-	"github.com/iotaledger/wasp/tools/wasp-cli/log"
+	"github.com/iotaledger/wasp/packages/log"
 )
 
 var metricsCmd = &cobra.Command{
